System Clock RCC Ready!
I've added the configurations needed to set the clock system to 180MHz (the highest possible) and configured some parameters to read data from the flash memory, also I turned on the prefetch and set the latency to read flash in 5 wait states according to operating voltage and clock frequency. We're ready to program other peripherals!